requirements.txt
llama-index>=0.10.0
openai>=1.0.0
faiss-cpu>=1.7.4
streamlit>=1.28.0
llama-index-llms-openai>=0.1.0
llama-index-llms-ollama>=0.1.0
llama-index-vector-stores-faiss>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
python-dotenv>=1.0.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
# pandas>=2.0.0
